requirements.txt
psutil
pywin32
numpy
pillow
opencv-python
tensorflow
keyboard